# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2.2.0
  description: Construction AI Platform — Phase 2 API (routes match FastAPI backend exactly)
servers:
  - url: /api
    description: Base API path (Python FastAPI backend)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserOut:
      type: object
      required: [id, email, role, is_active, created_at]
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: ["string", "null"]
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        must_change_password:
          type: boolean
          description: Phase 2 — Security & Authentication Hardening
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        remember_me:
          type: boolean
          default: false
          description: >-
            RC1 Phase 1 Sprint 1 — opt-in longer refresh-token sliding
            window for a trusted device (see RefreshTokenPolicy on the
            backend). Defaults to false.
        device:
          type: ["string", "null"]
          description: Optional client-supplied device label, recorded on the session.
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        full_name:
          type: ["string", "null"]
        role:
          type: string
          default: site_engineer
    TokenResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
        user:
          $ref: "#/components/schemas/UserOut"
    LoginResponse:
      description: >-
        RC1 Phase 1 Sprint 1 — response of POST /auth/login and POST
        /auth/refresh, both of which start/rotate a refresh-token session.
        NOT used by POST /auth/change-password, which still returns plain
        TokenResponse (it reissues an access token but never touches a
        session).
      allOf:
        - $ref: "#/components/schemas/TokenResponse"
        - type: object
          required: [refresh_token]
          properties:
            refresh_token:
              type: string
    RefreshInput:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
        device:
          type: ["string", "null"]
    LogoutInput:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    LogoutAllResponse:
      type: object
      required: [message, revoked_count]
      properties:
        message:
          type: string
        revoked_count:
          type: integer
    SessionOut:
      type: object
      required: [id, created_at, expires_at, remember_me]
      properties:
        id:
          type: integer
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: ["string", "null"]
          format: date-time
        expires_at:
          type: string
          format: date-time
        device:
          type: ["string", "null"]
        ip_address:
          type: ["string", "null"]
        user_agent:
          type: ["string", "null"]
        remember_me:
          type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Rotate a refresh token and mint a new access token
      description: >-
        Authenticates via the refresh_token in the request body, not a
        bearer access token — public like /auth/login. Reusing an
        already-rotated or already-revoked refresh token revokes its whole
        session (see backend app/core/session_security.py) and returns 401.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: Refresh successful — new access + refresh token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid, expired, or already-used (replayed) refresh token
  /v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Revoke the session the given refresh token belongs to
      description: >-
        Authenticates via the refresh_token in the request body. Idempotent
        by design — always returns 200, even for an unknown or
        already-revoked token, so it never leaks token validity.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutInput"
      responses:
        "200":
          description: Logged out (always, regardless of prior token state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
  /v1/auth/logout-all:
    post:
      operationId: logoutAllSessions
      tags: [auth]
      summary: Revoke every session for the authenticated caller
      responses:
        "200":
          description: All sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutAllResponse"
        "401":
          description: Unauthorized
  /v1/auth/sessions:
    get:
      operationId: listSessions
      tags: [auth]
      summary: List the authenticated caller's own active sessions
      responses:
        "200":
          description: Active sessions for the caller only — never cross-user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionOut"
        "401":
          description: Unauthorized
  /v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
        "409":
          description: Email already registered
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
        "401":
          description: Unauthorized
...